+++ server/CHANGELOG.md
@@ -1,3 +1,6 @@
+# v 1.1.0 (?)
+Changes in this release:
+
 # v 1.0.0 (2020-11-06)
 Changes in this release:
 - Added language server to client diagnostics notifications to highlight errors (#18)